requirements.txt
tensorflow==2.10.0
numpy
gdown
Pillow
streamlit

# python==3.10.0
# tensorflow_gpu==2.10.0